backend/requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
Flask-CORS==3.0.10
python-dotenv==0.19.0
requests==2.26.0
Werkzeug==2.0.1
gunicorn==20.1.0